Create UpsertPageImages
